services/cnn-extractor/requirements.txt
kafka-python>=2.0.2
numpy>=1.26.0
Pillow>=10.1.0
python-dotenv>=1.0.0
